Fix bug where environment variable DEBUG counterintuitively would not set to False.
The trouble with:
DEBUG = os.environ.get("DEBUG", True)

By default here DEBUG is a boolean. However, when you set the
environment variable DEBUG to "False" it is a string.  When a string
is evaluated as a boolean any string of lenght > 0 is true.

So instead we explicitly compare the string.lower() to "true". In which case it will evaluate as follows:
No setting for debug it will default to True
"True", "TRUE", "true" evaluate to True.
"", "False", "FALSE", "false" evaluate to False.